genus: generate clock constraints
